Moving new message types for house superweapons into new `Game Message` header file
